Move Update to its own header file
